seed phrase elements renamed
