+++ android/src/main/java/io/parity/signer/screens/keysets/restore/EnterSeedPhraseBox.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.screens.keysetdetails.backup
+package io.parity.signer.screens.keysets.restore
 
 import android.content.res.Configuration
 import androidx.compose.foundation.background
@@ -7,63 +7,59 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.google.accompanist.flowlayout.FlowMainAxisAlignment
 import com.google.accompanist.flowlayout.FlowRow
 import com.google.accompanist.flowlayout.SizeMode
 import io.parity.signer.R
-import io.parity.signer.domain.KeepScreenOn
 import io.parity.signer.domain.DisableScreenshots
+import io.parity.signer.domain.KeepScreenOn
+import io.parity.signer.screens.keysetdetails.backup.PhraseNumberStyle
+import io.parity.signer.screens.keysetdetails.backup.PhraseWordStyle
 import io.parity.signer.ui.theme.SignerNewTheme
 import io.parity.signer.ui.theme.fill6
 import io.parity.signer.ui.theme.textDisabled
 import io.parity.signer.ui.theme.textSecondary
 
-private val RobotoFontFamily = FontFamily(
-	Font(R.font.robotomono_regular, FontWeight.Medium),
-	Font(R.font.robotomono_bold, FontWeight.Bold),
-)
-internal val PhraseNumberStyle: TextStyle = TextStyle(
-	fontFamily = RobotoFontFamily,
-	fontWeight = FontWeight.Normal,
-	fontSize = 13.sp
-)
-internal val PhraseWordStyle: TextStyle = TextStyle(
-	fontFamily = RobotoFontFamily,
-	fontWeight = FontWeight.Bold,
-	fontSize = 13.sp
-)
 
 @Composable
-fun BackupPhraseBox(seedPhrase: String) {
+fun EnterSeedPhraseBox(phrase: String) {
 	val innerRound = dimensionResource(id = R.dimen.innerFramesCornerRadius)
 	val innerShape =
 		RoundedCornerShape(innerRound, innerRound, innerRound, innerRound)
 	FlowRow(
 		mainAxisSize = SizeMode.Expand,
 		mainAxisAlignment = FlowMainAxisAlignment.SpaceBetween,
+		crossAxisSpacing = 4.dp,
 		modifier = Modifier
 			.padding(horizontal = 16.dp)
 			.padding(top = 8.dp, bottom = 16.dp)
 			.background(MaterialTheme.colors.fill6, innerShape)
 			.padding(16.dp),
 	) {
-		val words = seedPhrase.split(" ")
+
+		val editedWord = phrase.substringAfterLast(' ', "")
+		val words =
+			phrase.split(' ').filter() { it.isNotEmpty() }//todo fix last word?
+
 		for (index in 0..words.lastIndex) {
-			BackupPhraseItem(index = index + 1, word = words[index])
+			EnterSeedPhraseWord(index = index + 1, word = words[index])
 		}
+		BasicTextField(
+			modifier = Modifier
+				.defaultMinSize(minWidth = 100.dp, minHeight = 24.dp)
+				.padding(vertical = 8.dp, horizontal = 12.dp),
+			value = editedWord,
+			onValueChange = {},//todo dmitry
+		)
 	}
 
 	DisableScreenshots()
@@ -72,17 +68,18 @@ fun BackupPhraseBox(seedPhrase: String) {
 
 
 @Composable
-private fun BackupPhraseItem(index: Int, word: String) {
-	Row(Modifier
-		.defaultMinSize(minWidth = 100.dp, minHeight = 24.dp)
-		.padding(vertical = 2.dp)
+private fun EnterSeedPhraseWord(index: Int, word: String) {
+	Row(
+		Modifier
+			.background(MaterialTheme.colors.fill6, RoundedCornerShape(16.dp))
+			.defaultMinSize(minWidth = 100.dp, minHeight = 24.dp)
+			.padding(vertical = 8.dp, horizontal = 12.dp)
 	) {
 		Text(
 			text = index.toString(),
 			color = MaterialTheme.colors.textDisabled,
 			style = PhraseNumberStyle,
 			textAlign = TextAlign.End,
-			modifier = Modifier.defaultMinSize(minWidth = 16.dp)
 		)
 		Spacer(Modifier.padding(start = 6.dp))
 		Text(
@@ -92,7 +89,6 @@ private fun BackupPhraseItem(index: Int, word: String) {
 		)
 	}
 }
-
 
 @Preview(
 	name = "light", group = "general", uiMode = Configuration.UI_MODE_NIGHT_NO,
@@ -104,8 +100,25 @@ private fun BackupPhraseItem(index: Int, word: String) {
 	showBackground = true, backgroundColor = 0xFF000000,
 )
 @Composable
-private fun PreviewBackupPhraseBox() {
+private fun PreviewSeedPhraseRestoreComponentFinished() {
 	SignerNewTheme {
-		BackupPhraseBox("some workds used for secret veryverylong special long text here to see how printed")
+		EnterSeedPhraseBox("some workds used for secret veryverylong special long text here to see how printed ")
 	}
 }
+
+@Preview(
+	name = "light", group = "general", uiMode = Configuration.UI_MODE_NIGHT_NO,
+	showBackground = true, backgroundColor = 0xFFFFFFFF,
+)
+@Preview(
+	name = "dark", group = "general",
+	uiMode = Configuration.UI_MODE_NIGHT_YES,
+	showBackground = true, backgroundColor = 0xFF000000,
+)
+@Composable
+private fun PreviewSeedPhraseRestoreComponentInProgress() {
+	SignerNewTheme {
+		EnterSeedPhraseBox("some workds used for secret veryverylong special long text here to see how printed")
+	}
+}
+
